feat: articles use case testing

+++ backend/src/core/article/service/DeleteArticle.ts
@@ -1,5 +1,6 @@
 import UseCase from "@/core/shared/UseCase"
 import ArticleRepository from "./ArticleRepository"
+import Erros from "@/core/shared/Erros"
 
 export interface In {
     articleId: string
@@ -9,6 +10,10 @@ export default class DeleteArticle implements UseCase<In, void> {
     constructor(private repository: ArticleRepository) {}
 
     async execute(In: In): Promise<void> {
+        const article = await this.repository.getArticleById(In.articleId)
+
+        if(!article) throw new Error(Erros.ARTICLE_NOT_EXISTS)
+
         await this.repository.deleteArticle(In.articleId)
     }
 }
